editoast: refacto timetable modules
